run.c内で速度をmm/s→Hzで処理するように変更。　それに伴い、PIDパラメータ変更 探索走行、最短走行にslalom追加。　引数でslalomのパラメータ指定できるようにする予定。 探索走行時の壁読み位置変更。車体中央が区間の境目に来た時に読む。 次回、slalomの前後の空走距離指定処理追加予定。

+++ Core/Inc/motor.h
@@ -9,6 +9,8 @@
 #define INC_MOTOR_H_
 
 #include "main.h"
+#include "arm_math.h"
+#include "arm_const_structs.h"
 
 
 
@@ -23,8 +25,8 @@
 
 
 
-uint32_t SPEEDtoHz(uint32_t speed);
-uint32_t HztoSPEED(uint32_t Hz);
+float32_t SPEEDtoHz(float32_t speed);
+float32_t HztoSPEED(float32_t Hz);
 void mortor_direction(uint8_t motor,uint8_t direction);
 void mortor_stop(void);
 #endif /* INC_MOTOR_H_ */
